rename vc
